ml-service/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.2",
    "ollama>=0.6.1",
    "pillow>=12.1.0",
    "pydantic-settings>=2.12.0",
    "pytesseract>=0.3.13",
    "python-telegram-bot>=22.6",
    "uvicorn[standard]>=0.40.0",
]
